+++ navigate/core/node_reference.py
@@ -3,8 +3,6 @@
 
 from __future__ import annotations
 
-import numpy as np
-
 from navigate.core.node_type import TypeCheckMixin
 
 
@@ -13,18 +11,6 @@ class NodeReference(TypeCheckMixin):
         super().__init__(type_)
 
         self.name = name
-
-        # string indicating the file and line in
-        # the deck where the node is referenced
-        self.reference_location = ""
-
-        # this attribute is only relevant for calculators
-        # they are set during the call to assign_value if
-        # bounds on the attribute are included in the call
-        self.internal_bounds = (-np.inf, np.inf)
-
-    def set_internal_bounds(self, lower, upper):
-        self.internal_bounds = (lower, upper)
 
     def __repr__(self):
         return f'{self.type}("{self.name}")'